feat: add download function
