--- .husky/pre-commit
@@ -1,1 +1,0 @@
-
