+++ src/fnHelpers/fnHelpers.ts
@@ -1,0 +1,20 @@
+import { IPokeItem } from "../types";
+
+export const convertToCSV = (data: IPokeItem[]): string => {
+  if (!data.length) {
+    return '';
+  }
+
+  const csvRows: string[] = [];
+  const headers = Object.keys(data[0]);
+  csvRows.push(headers.join(','));
+  for (const row of data) {
+    const values = headers.map(header => {
+      const escaped = ('' + row[header as keyof IPokeItem]).replace(/"/g, '\\"');
+      return `"${escaped}"`;
+    });
+    csvRows.push(values.join(','));
+  }
+
+  return csvRows.join('\n');
+};
