+++ src/store/index.ts
@@ -2,20 +2,17 @@ import { configureStore } from '@reduxjs/toolkit';
 import PokeSlice from './PokeSlice';
 import PageSlice from './PageSlice';
 import apiSlice from './ApiSlice';
-/* import DownloadApiSlice from './DownloadApiSlice'; */
 
 
 const store = configureStore({
   reducer: {
     PokeStore: PokeSlice,
     PageStore: PageSlice,
-    [apiSlice.reducerPath]: apiSlice.reducer,
-    /* [DownloadApiSlice.reducerPath]: DownloadApiSlice.reducer */
+    [apiSlice.reducerPath]: apiSlice.reducer
   },
   middleware: (getDefaultMiddleware) =>
     getDefaultMiddleware()
       .concat(apiSlice.middleware)
-      /* .concat(DownloadApiSlice.middleware) */,
 });
